requirements.txt
aiohttp==3.6.*
aiofiles==0.4.0
environs-9.5.0